requirements.txt
# requires Python 3.11+
duckdb>=1.1.0
pydantic>=2.7.0
pyyaml>=6.0.1
snowflake-connector-python>=3.14.0
databricks-sql-connector>=3.1.0